requirements.txt
aniso8601==8.0.0
attrs==19.3.0
autopep8==1.5
click==7.1.1
entrypoints==0.3
flake8==3.7.9
flask-restplus==0.13.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
wcwidth==0.1.8
Werkzeug==0.16.1
zipp==3.1.0
tensorflow==1.15.0
Keras==2.2.4